move vendor javascript files to vendors folder
