+++ src/static/js/rjquery.js
@@ -1,5 +1,5 @@
 'use strict';
 // Provides a require'able version of jQuery without leaking $ and jQuery;
-window.$ = require('./jquery');
+window.$ = require('./vendors/jquery');
 const jq = window.$.noConflict(true);
 exports.jQuery = exports.$ = jq;
