+++ src/static/js/timeslider.js
@@ -24,7 +24,7 @@
 
 // These jQuery things should create local references, but for now `require()`
 // assigns to the global `$` and augments it with plugins.
-require('./jquery');
+require('./vendors/jquery');
 
 const Cookies = require('./pad_utils').Cookies;
 const randomString = require('./pad_utils').randomString;
