+++ src/static/js/vendors/browser.js
