+++ src/static/js/vendors/farbtastic.js
