+++ src/templates/admin/settings.html
@@ -4,14 +4,14 @@
     <title data-l10n-id="admin_settings.page-title">Settings - Etherpad</title>
     <meta name="viewport" content="width=device-width">
     <link rel="stylesheet" href="../static/css/admin.css">
-    <script src="../static/js/jquery.js"></script>
+    <script src="../static/js/vendors/jquery.js"></script>
     <script src="../socket.io/socket.io.js"></script>
     <script src="../static/js/socketio.js"></script>
     <script src="../static/js/admin/minify.json.js"></script>
     <script src="../static/js/admin/settings.js"></script>
     <script src="../static/js/admin/jquery.autosize.js"></script>
     <link rel="localizations" type="application/l10n+json" href="../locales.json" />
-    <script src="../static/js/html10n.js"></script>
+    <script src="../static/js/vendors/html10n.js"></script>
     <script src="../static/js/l10n.js"></script>
   </head>
   <body>
