+++ src/templates/index.html
@@ -11,9 +11,9 @@
         <link rel="shortcut icon" href="<%=settings.favicon%>">
 
         <link rel="localizations" type="application/l10n+json" href="locales.json">
-        <script type="text/javascript" src="static/js/html10n.js?v=<%=settings.randomVersionString%>"></script>
+        <script type="text/javascript" src="static/js/vendors/html10n.js?v=<%=settings.randomVersionString%>"></script>
         <script type="text/javascript" src="static/js/l10n.js?v=<%=settings.randomVersionString%>"></script>
-        <script src="static/js/jquery.js"></script>
+        <script src="static/js/vendors/jquery.js"></script>
         <script src="static/js/index.js"></script>
 
         <style>
