+++ src/tests/frontend/index.html
@@ -9,8 +9,8 @@
     <div id="mocha"></div>
     <div id="iframe-container"></div>
 
-    <script src="/static/js/jquery.js"></script>
-    <script src="/static/js/browser.js"></script>
+    <script src="/static/js/vendors/jquery.js"></script>
+    <script src="/static/js/vendors/browser.js"></script>
     <script src="lib/underscore.js"></script>
 
     <script src="lib/mocha.js"></script>
@@ -20,7 +20,7 @@
     <script src="helper.js"></script>
     <script src="helper/methods.js"></script>
     <script src="helper/ui.js"></script>
-    
+
     <script src="specs_list.js"></script>
     <script src="runner.js"></script>
 </html>
